Create Given an array A[n] of n numbers, create an output array such that out[i] is equal to the product of all the elements of A[n] except A[i]. Ex. out[4] assuming n >= 5 will be the product of the numbers from A[0] through A[3] and the numbers from A[5] through A[n - 1]. print the out array.py
Given an array A[n] of n numbers, create an output array such that out[i] is equal to the product of all the elements of A[n] except A[i]. Ex. out[4] assuming n >= 5 will be the product of the numbers from A[0] through A[3] and the numbers from A[5] through A[n - 1]. print the out array.